fixed build errors from moving types around and minor update to support it
