+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "goobs-repo",
-  "version": "0.7.15",
+  "version": "0.7.16",
   "license": "MIT",
   "main": "./src/index.ts",
   "types": "./src/index.ts",
